requirements.txt
boto3>=1.34.0
psycopg2-binary>=2.9.9
slack-sdk>=3.24.1
python-dotenv>=1.0.1
structlog>=23.3.0
PyYAML>=6.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
protobuf==3.20.3  # Fixes compatibility with FAISS
scikit-learn==1.1.3 # behaviour param still supported here
pandas>=2.0.0
numpy==1.23.5
transformers>=4.39.0
torch>=1.13.1
accelerate>=0.28.0
bitsandbytes>=0.41.1
huggingface_hub>=0.23.2